pyproject.toml
[project]
name = "label-verification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pypdf>=6.7.1",
    "pytest>=9.0.2",
    "python-doctr[torch]>=1.0.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "uvicorn>=0.40.0",
]
